requirements.txt
sentence-transformers
numpy
faiss-cpu
black
pytest
